+++ packages/web/src/controls/SearchInput.tsx
@@ -1,9 +1,8 @@
 import React, { forwardRef, memo, useCallback, useMemo, useRef } from 'react';
 import { useMergeRefs } from '@coinbase/cds-common/hooks/useMergeRefs';
 import type { IconName } from '@coinbase/cds-common/types';
-import { css } from '@linaria/core';
 
-import { cx } from '../cx';
+import { useComponentConfig } from '../hooks/useComponentConfig';
 import { Box } from '../layout/Box';
 
 import { InputIcon } from './InputIcon';
@@ -14,14 +13,6 @@ export const scales = {
   regular: 56,
   compact: 40,
 };
-
-const baseCss = css`
-  height: ${scales.regular}px;
-`;
-
-const compactCss = css`
-  height: ${scales.compact}px;
-`;
 
 export type SearchInputBaseProps = Pick<
   TextInputBaseProps,
@@ -89,8 +80,9 @@ export type SearchInputProps = SearchInputBaseProps &
   };
 
 export const SearchInput = memo(
-  forwardRef(function SearchInput(
-    {
+  forwardRef((_props: SearchInputProps, ref: React.ForwardedRef<HTMLInputElement>) => {
+    const mergedProps = useComponentConfig('SearchInput', _props);
+    const {
       onChange,
       onClear,
       onChangeText,
@@ -105,10 +97,9 @@ export const SearchInput = memo(
       startIconAccessibilityLabel = 'Back',
       clearIconAccessibilityLabel = 'Clear search query',
       borderRadius = 1000,
+      height = compact ? scales.compact : scales.regular,
       ...props
-    }: SearchInputProps,
-    ref: React.ForwardedRef<HTMLInputElement>,
-  ) {
+    } = mergedProps;
     const internalRef = useRef<HTMLInputElement>(null);
     const refs = useMergeRefs(ref, internalRef);
 
@@ -148,7 +139,6 @@ export const SearchInput = memo(
       <TextInput
         ref={refs}
         borderRadius={borderRadius}
-        className={cx(baseCss, compact && compactCss)}
         end={
           end ??
           (!!value && !hideEndIcon && (
@@ -162,6 +152,7 @@ export const SearchInput = memo(
             </Box>
           ))
         }
+        height={height}
         onChange={handleOnChange}
         onKeyUp={handleOnKeyUp}
         role="searchbox"